Safari-611.1.10/LayoutTests/fast/dom/microtask-promise-mutation-observer-order-expected.txt

Test that promises and mutation observers are both delivered from the same microtask queue.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Expected result: promise,mutate,timeout
Actual result: promise,mutate,timeout
PASS actualResult is expectedResult

Expected result: mutate,promise,timeout
Actual result: mutate,promise,timeout
PASS actualResult is expectedResult

PASS successfullyParsed is true

TEST COMPLETE
